Las facturas de mayorista pueden incluir unidades fueras de bultos, que pueden ser rechazadas. Se debe dividir las unidades por bultos traidas en el json

+++ app/src/main/java/com/dynamicsoftware/pocho/logistica/Vista/EntregaParcial/FacturasExpandableListAdapter.java
@@ -145,7 +145,7 @@ public class FacturasExpandableListAdapter extends BaseExpandableListAdapter
         TextView tvImporteFinal = (TextView) convertView.findViewById(R.id.tvImporteFinal);
         tvArticulo.setText(itemFactura.getArticulo());
         tvDescripcion.setText(itemFactura.getDescripcion());
-        int cantidad = itemFactura.getCantidad();
+        double cantidad = itemFactura.getCantidad();
         tvCantidad.setText(String.valueOf(cantidad));
         if (cantidad < 0)
         {
